Fix jsdocs and move file upload custom hook into its own file
